refactor: constants 이동
